requirements.txt
llama-cpp-python==0.2.24
openai
streamlit
PyPDF2
sentence-transformers
faiss-cpu
langchain
numpy

